flask/requirements.txt
appdirs==1.4.4
certifi==2019.11.28
click==7.1.2
distlib==0.3.0
dnspython==1.16.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
uWSGI==2.0.18
virtualenv==20.0.21
Werkzeug==1.0.1
WTForms==2.3.1
